==> code/01_behavior_analysis_n43.R <==
# ____________________________________________________________________________________________
# AWE IS CHARACTERIZED AS AN AMBIVALENT AFFECT IN THE HUMAN BEHAVIOR AND CORTEX
# _________________________________
# 01 behavioral_analysis_n43.R
# _________________________________
# written by
# /
# Jinwoo Lee (SNU Connectome Lab)
# e-mail:  adem1997@snu.ac.kr
# website: jinwoo-lee.com
# _________________________________
# July, 2025
# ____________________________________________________________________________________________
# [MATERIALS GENERATED BY THIS SCRIPT] Fig 2, Fig 3, STable 1, STable 3

# loading required packages
library(readxl)
library(ltm)
library(ggplot2)
library(lsr)
library(ggpubr)
library(lmerTest)
library(DHARMa)
library(h2o)
library(brunnermunzel)
library(bayestestR)

### ________________________________________________________ ####
### @@@ I. DATA LOADING & FORMATTING @@@ ####
sub.list.behav <- read.csv("../dataset/sub-list-full_n43.csv", header = FALSE)
sub.list.behav <- sub.list.behav[[1]]

clip.list <- c("SP", "CI", "MO", "PA")

## > I-1. Pre-experiment Survey Dataset ####
df.pre <- read_excel("../dataset/behavior/pre-experiment-survey_n43.xlsx", "pre-experiment-survey_n43")
df.pre <- df.pre[df.pre$subjectkey %in% sub.list.behav, 
                 c("subjectkey", "sex_M", "age", "panas_p_total", "panas_n_total", 
                   "dpes_awe_01", "dpes_awe_02", "dpes_awe_03", 
                   "dpes_awe_04", "dpes_awe_05", "dpes_awe_06", "dpes_awe_total")]

boxplot(df.pre$panas_n_total) # there are not any outliers
print(paste0("[Negative mood] mean = ", round(mean(df.pre$panas_n_total), 3), " / SD = ", round(sd(df.pre$panas_n_total), 3)))

boxplot(df.pre$panas_p_total) # there are not any outliers
print(paste0("[Positive mood] mean = ", round(mean(df.pre$panas_p_total), 3), " / SD = ", round(sd(df.pre$panas_p_total), 3)))

# check the psychometric reliability of translated DPES items
print(paste0("[DPES-AWE] mean = ", round(mean(df.pre$dpes_awe_total), 3), " / SD = ", round(sd(df.pre$dpes_awe_total), 3)))
dpes.cron.alpha <- cronbach.alpha(df.pre[, c("dpes_awe_01", "dpes_awe_02", "dpes_awe_03", "dpes_awe_04", "dpes_awe_05", "dpes_awe_06")])[[1]]

## > I-2. Post-experiment Self-report Dataset ####
df.SP <- read_excel("../dataset/behavior/report_SP_n43.xlsx", "report_SP_n43")
df.SP <- merge(df.pre, df.SP, by = 'subjectkey')
df.CI <- read_excel("../dataset/behavior/report_CI_n43.xlsx", "report_CI_n43")
df.CI <- merge(df.pre, df.CI, by = 'subjectkey')
df.MO <- read_excel("../dataset/behavior/report_MO_n43.xlsx", "report_MO_n43")
df.MO <- merge(df.pre, df.MO, by = 'subjectkey')
df.PA <- read_excel("../dataset/behavior/report_PA_n43.xlsx", "report_PA_n43")
df.PA <- merge(df.pre, df.PA, by = 'subjectkey')

rm(df.pre) # clean the environment

# transplant keypress features into clip datasets
for (sub_idx in c(1:length(sub.list.behav))) {
  for (clip_idx in c(1:length(clip.list))) {
    
    # search preprocessed EEG-Valence paired dataset
    tmp_sub <- sub.list.behav[sub_idx]
    tmp_clip <- clip.list[clip_idx]
    tmp_filename <- paste0(tmp_sub, "_", tmp_clip, "-norm-CEBRA.csv")
    tmp_dir <- paste0("../dataset/eeg/pped/", tmp_sub, "/CEBRA_input/", tmp_filename)
    tmp_file <- read.csv(tmp_dir, header = TRUE)
    
    # get the subject's keypress vector
    tmp_file$type <- as.factor(tmp_file$type)
    type_table <- as.data.frame(table(tmp_file$type))
    len_none <- type_table[type_table$Var1 == '0', 'Freq'] / length(tmp_file$type)
    if (length(len_none) == 0) {len_none <- 0}
    len_pos <- type_table[type_table$Var1 == '1', 'Freq'] / length(tmp_file$type)
    if (length(len_pos) == 0) {len_pos <- 0}
    len_mix <- type_table[type_table$Var1 == '2', 'Freq'] / length(tmp_file$type)
    if (length(len_mix) == 0) {len_mix <- 0}
    len_neg <- type_table[type_table$Var1 == '3', 'Freq'] / length(tmp_file$type)
    if (length(len_neg) == 0) {len_neg <- 0}
    
    # transplant one's keypress features into clip dataset
    tmp_target <- get(paste0('df.', tmp_clip))
    tmp_target[sub_idx, "len_none"] <- len_none
    tmp_target[sub_idx, "len_pos"] <- len_pos
    tmp_target[sub_idx, "len_mix"] <- len_mix
    tmp_target[sub_idx, "len_neg"] <- len_neg
    
    # update clip dataset
    assign(paste0('df.', tmp_clip), tmp_target)
  }
  
  # verbose
  print(paste0("[", sub_idx, "/", length(sub.list.behav), "] The keypress features were copied successfully to all clip datasets!"))
}

# clean the environment
rm(tmp_file, tmp_target, clip_idx, tmp_clip, tmp_dir, tmp_filename, 
   tmp_sub, len_mix, len_neg, len_none, len_pos, sub_idx, type_table)

# merge the dataset
data <- rbind(df.SP, df.CI)
data <- rbind(data, df.MO)
data <- rbind(data, df.PA)

# define the variable type
data$subjectkey <- as.factor(data$subjectkey)
data$clip <- as.factor(data$clip)
data$order <- as.factor(data$order)

df.SP <- data[data$clip == "SP", ]
df.CI <- data[data$clip == "CI", ]
df.MO <- data[data$clip == "MO", ]
df.PA <- data[data$clip == "PA", ]

# check the psychometric reliability of AWES scale
awes.cron.alpha <- cronbach.alpha(data[, c(15:44)])[[1]]


### ________________________________________________________ ####
### @@@ II. KEYPRESS DYNAMICS VISUALIZATION [FIG 2A] @@@ ####
df.key.SP <- data.frame(matrix(nrow = 0, ncol = 3))
df.key.CI <- data.frame(matrix(nrow = 0, ncol = 3))
df.key.MO <- data.frame(matrix(nrow = 0, ncol = 3))
df.key.PA <- data.frame(matrix(nrow = 0, ncol = 3))

## > II-1. Loading the STFT-processed Valence Keypress ####
for (clip_idx in c(1:length(clip.list))){
  # define the clip iteratively
  tmp_clip <- clip.list[clip_idx]
  
  # load the dataframe to fill in
  tmp_target_df <- get(paste0("df.key.", tmp_clip))
  
  # transplant each subject's keypress timeseries into the target dataframe
  for (sub_idx in c(1:length(sub.list.behav))){
    tmp_sub <- sub.list.behav[sub_idx]
    
    # load the keypress timeseries information
    tmp_filename <- paste0(tmp_sub, "_", tmp_clip, "-STFT-CEBRA.csv")
    tmp_dir <- paste0("../dataset/eeg/pped/", tmp_sub, "/CEBRA_input/")
    tmp_file <- read.csv(paste0(tmp_dir, tmp_filename), header = TRUE)
    tmp_keypress_series <- tmp_file$mode
    
    # iteratively archive keypress of each timepoint 
    for (sec in c(1:length(tmp_keypress_series))){
      tmp_time <- sprintf("sec%03d", sec)
      tmp_key <- tmp_keypress_series[sec]
      
      tmp_target_df <- rbind(tmp_target_df, c(tmp_sub, tmp_time, tmp_key))
    }
  }
  
  # rename columns 
  colnames(tmp_target_df) <- c("subjectkey", "timepoint", "key")
  
  # make each column factor
  tmp_target_df$subjectkey <- as.factor(tmp_target_df$subjectkey)
  tmp_target_df$timepoint <- as.factor(tmp_target_df$timepoint)
  tmp_target_df$key <- factor(tmp_target_df$key, levels = c("0", "1", "2", "3"))
  
  # update
  assign(paste0('df.key.', tmp_clip), tmp_target_df)
  
  # verbosa
  print(paste0("[", clip_idx, "/", length(clip.list), "] All subjects' keypress timeseries were successfully transplanted to ", 
               clip.list[clip_idx], " data-frame!"))
}

# clean the environment
rm(tmp_file, tmp_target_df, clip_idx, tmp_clip, tmp_dir, tmp_filename, tmp_key, 
   tmp_keypress_series, tmp_sub, tmp_time, sec, sub_idx)

## > II-2. Visualization ####
fig2a.SP <- ggplot(df.key.SP, aes(x = timepoint, y = subjectkey, fill = key)) +
  scale_fill_manual(values = c("#E0E0E0", "#EF9A9A", "#CE93D8", "#9FA8DA")) + # neutral, positive, mixed, negative
  theme(axis.text.x = element_blank(),
        axis.ticks.x = element_blank()) +
  xlab("") +
  ylab("") +
  geom_tile()

fig2a.CI <- ggplot(df.key.CI, aes(x = timepoint, y = subjectkey, fill = key)) +
  scale_fill_manual(values = c("#E0E0E0", "#EF9A9A", "#CE93D8", "#9FA8DA")) + # neutral, positive, mixed, negative
  theme(axis.text.x = element_blank(),
        axis.ticks.x = element_blank()) +
  xlab("") +
  ylab("") +
  geom_tile()

fig2a.MO <- ggplot(df.key.MO, aes(x = timepoint, y = subjectkey, fill = key)) +
  scale_fill_manual(values = c("#E0E0E0", "#EF9A9A", "#CE93D8", "#9FA8DA")) + # neutral, positive, mixed, negative
  theme(axis.text.x = element_blank(),
        axis.ticks.x = element_blank()) +
  xlab("") +
  ylab("") +
  geom_tile()

fig2a.PA <- ggplot(df.key.PA, aes(x = timepoint, y = subjectkey, fill = key)) +
  scale_fill_manual(values = c("#E0E0E0", "#EF9A9A", "#CE93D8", "#9FA8DA")) + # neutral, positive, mixed, negative
  theme(axis.text.x = element_blank(),
        axis.ticks.x = element_blank()) +
  xlab("") +
  ylab("") +
  geom_tile()


### ________________________________________________________ ####
### @@@ III. AWES RATING COMPARISON [FIG 2B] @@@ ####
## > III-1. Testing Conceptual Model ####
data$awes_vastness <- (data$awes_06 + data$awes_07 + data$awes_08 + data$awes_09 + data$awes_10) / 5
data$awes_accommodation <- (data$awes_26 + data$awes_27 + data$awes_28 + data$awes_29 + data$awes_30) / 5

pos_lmm <- lmer(esg_pos ~ data$awes_vastness + data$awes_accommodation + (1|subjectkey), data = data)
summary(pos_lmm)

neg_lmm <- lmer(esg_neg ~ data$awes_vastness + data$awes_accommodation + (1|subjectkey), data = data)
summary(neg_lmm)

amb_int_lmm <- lmer(esg_eci ~ data$motion_sickness + (1|subjectkey) + (1|clip), data = data)
summary(amb_int_lmm)

amb_dur_lmm <- lmer(len_mix ~ data$motion_sickness + (1|subjectkey) + (1|clip), data = data)
summary(amb_dur_lmm)


## > III-2. Customized T-Test Function ####
run_t_tests <- function(var_name) {
  result_df <- data.frame(matrix(ncol = 7, nrow = 0))
  colnames(result_df) <- c("test", "t", "df", "ci_lower", "ci_upper", "p_value", "cohens_d")
  
  for (current_clip in clip.list[-length(clip.list)]) {
    current_df <- get(paste0("df.", current_clip))
    t_test <- t.test(current_df[[var_name]], df.PA[[var_name]], alternative = "two.sided", paired = TRUE)
    
    result_df <- rbind(result_df, data.frame(
      test = paste0(current_clip, " - PA"),
      t = round(t_test$statistic[[1]], 3),
      df = round(t_test$parameter[[1]], 3),
      ci_lower = round(t_test$conf.int[1], 3),
      ci_upper = round(t_test$conf.int[2], 3),
      p_value = t_test$p.value,
      cohens_d = round(cohensD(current_df[[var_name]], df.PA[[var_name]], method = "paired"), 3)
    ))
  }
  
  result_df$p_value_fdr <- p.adjust(result_df$p_value, "fdr")
  return(result_df)
}

## > III-3. Statistical Testing #### 
tt.awes <- run_t_tests("awes_avg")

## > III-4. Visualization ####
fig2b <- ggplot(data, aes(x = clip, y = awes_avg, fill = clip)) +
  geom_violin(aes(fill = clip), alpha = 0.5, color = "white") +
  scale_x_discrete(limits = c("SP", "CI", "MO", "PA")) + 
  scale_y_continuous(limits = c(0.0, 6.0)) + 
  scale_fill_manual(values = c("#E1BEE7", "#E1BEE7", "#BDBDBD", "#E1BEE7")) +
  geom_point(aes(group = clip), position = position_nudge(), alpha = 0.25, shape = 16) +
  geom_line(aes(group = subjectkey), position = position_nudge(), color = "#616161", alpha = 0.15) +
  theme_bw() + 
  theme(panel.border = element_blank(), panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(), axis.line = element_line(colour = "black")) +
  ggtitle("AWES") 


### ________________________________________________________ ####
### @@@ IV. AFFECT COMPARISON [FIG 2C & STABLE 3] @@@ ####
## > IV-1. Statistical Testing ####
tt.int_amb <- run_t_tests("esg_eci") ## >> IV-1-a. intensity of ambivalence ####
tt.int_amb$condition <- "int_amb"
tt.dur_amb <- run_t_tests("len_mix") ## >> IV-1-b. duration of ambivalence ####
tt.dur_amb$condition <- "dur_amb"
tt.int_pos <- run_t_tests("esg_pos") ## >> IV-1-c. intensity of positivity ####
tt.int_pos$condition <- "int_pos"
tt.dur_pos <- run_t_tests("len_pos") ## >> IV-1-d. duration of positivity ####
tt.dur_pos$condition <- "dur_pos"
tt.int_neg <- run_t_tests("esg_neg") ## >> IV-1-e. intensity of negativity ####
tt.int_neg$condition <- "int_neg"
tt.dur_neg <- run_t_tests("len_neg") ## >> IV-1-f. duration of negativity ####
tt.dur_neg$condition <- "dur_neg"
tt.arousal <- run_t_tests("arousal") ## >> IV-1-g. arousal ####
tt.arousal$condition <- "arousal"

tt.affect.total <- rbind(tt.int_amb, tt.dur_amb)
tt.affect.total <- rbind(tt.affect.total, tt.int_pos)
tt.affect.total <- rbind(tt.affect.total, tt.dur_pos)
tt.affect.total <- rbind(tt.affect.total, tt.int_neg)
tt.affect.total <- rbind(tt.affect.total, tt.dur_neg)
tt.affect.total <- rbind(tt.affect.total, tt.arousal)

write.csv(tt.affect.total, "../results/behavioral_results/ttest_affect.csv") # STable 3

## > IV-2. Visualization ####
# >> IV-2-a. duration of ambivalence ####
fig2c.dur <- ggplot(data, aes(x = clip, y = len_mix, fill = clip)) +
  geom_violin(aes(fill = clip), alpha = 0.5, color = "white") +
  scale_x_discrete(limits = c("SP", "CI", "MO", "PA")) + 
  scale_fill_manual(values = c("#E1BEE7", "#E1BEE7", "#BDBDBD", "#E1BEE7")) +
  geom_point(aes(group = clip), position = position_nudge(), alpha = 0.25, shape = 16) +
  geom_line(aes(group = subjectkey), position = position_nudge(), color = "#616161", alpha = 0.15) +
  theme_bw() + 
  theme(panel.border = element_blank(), panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(), axis.line = element_line(colour = "black")) +
  ggtitle("Duration of ambivalence") 

# >> IV-2-b. intensity of ambivalence ####
data$esg_eci <- as.factor(data$esg_eci)
fig2c.int.df <- data[, c("subjectkey", "clip", "esg_eci")]
fig2c.int.table <- table(fig2c.int.df$esg_eci, fig2c.int.df$clip)
fig2c.int.table <- as.data.frame.matrix(fig2c.int.table)

fig2c.int <- ggballoonplot(fig2c.int.table, color = "white") +
  scale_x_discrete(limits = c("SP", "CI", "MO", "PA")) + 
  scale_y_discrete(limits = c(1, 2, 3, 4))

data$esg_eci <- as.numeric(data$esg_eci) # for the next regression analysis


### ________________________________________________________ ####
### @@@ V. UNIVARIATE LINEAR MODELING [FIG 3A] @@@ ####
## > V-1. Statistical Testing ####
predictor.list <- c("sex_M", "age", "panas_p_total", "panas_n_total", "dpes_awe_total", 
                    "esg_pos", "esg_neg", "esg_eci", "arousal", "motion_sickness", 
                    "len_none", "len_pos", "len_neg", "len_mix")

res.uni.lmm <- data.frame(matrix(nrow = 0, ncol = 9))

for (idx in c(1:length(predictor.list))) {
  tmp_predictor <- predictor.list[idx]
  tmp_lmm <- lmer(awes_avg ~ data[, tmp_predictor] + (1|subjectkey) + (1|clip), data = data)
  tmp_lmm_info <- summary(tmp_lmm)
  
  coef <- tmp_lmm_info$coefficients[[2]]
  se <- tmp_lmm_info$coefficients[[4]]
  ci_lower <- coef - 1.96*se
  ci_upper <- coef + 1.96*se
  df <- tmp_lmm_info$coefficients[[6]]
  t <- tmp_lmm_info$coefficients[[8]]
  p <- tmp_lmm_info$coefficients[[10]]
  
  tmp_simulated_residuals <- simulateResiduals(fittedModel = tmp_lmm)
  ks_test_result <- testUniformity(tmp_simulated_residuals)
  
  ks_p <- ks_test_result$p.value
  
  res.uni.lmm <- rbind(res.uni.lmm, c(tmp_predictor, ks_p, coef, se, ci_lower, ci_upper, df, t, p))
}

rm(idx, tmp_predictor, tmp_lmm, tmp_lmm_info, coef, se, ci_lower, ci_upper, df, t, p,
   tmp_simulated_residuals, ks_test_result, ks_p) # clean the environment

colnames(res.uni.lmm) <- c("predictor", "ks_test_p_value", "coefficient", "se", 
                           "ci_lower", "ci_upper", "df", "t_value", "p_value")

res.uni.lmm$predictor <- as.factor(res.uni.lmm$predictor)
for (i in c(2:ncol(res.uni.lmm))) {res.uni.lmm[, i] <- as.numeric(res.uni.lmm[, i])}

# save the results
write.csv(res.uni.lmm, "../results/behavioral_results/univariate_modeling_stats.csv")

## > V-2. Testing Confounding Effects of Motion Sickness ####
ms_awes_null <- lmer(awes_avg ~ 1 + (1|subjectkey) + (1|clip), data = data)
ms_awes_full <- lmer(awes_avg ~ data[, "motion_sickness"] + (1|subjectkey) + (1|clip), data = data)
summary(ms_awes_full)
bayesfactor_models(ms_awes_full, ms_awes_null)

ms_lenmix_null <- lmer(len_mix ~ 1 + (1|subjectkey) + (1|clip), data = data)
ms_lenmix_full <- lmer(len_mix ~ data[, "motion_sickness"] + (1|subjectkey) + (1|clip), data = data)
summary(ms_lenmix_full)
bayesfactor_models(ms_lenmix_full, ms_lenmix_null)

ms_intmix_null <- lmer(esg_eci ~ 1 + (1|subjectkey) + (1|clip), data = data)
ms_intmix_full <- lmer(esg_eci ~ data[, "motion_sickness"] + (1|subjectkey) + (1|clip), data = data)
summary(ms_intmix_full)
bayesfactor_models(ms_intmix_full, ms_intmix_null)

## > V-3. Visualization ####
fig3a <- ggplot(res.uni.lmm, aes(x = coefficient, y = predictor)) +
  geom_point() +
  scale_y_discrete(limits = c("motion_sickness", "arousal", "esg_neg", "esg_eci", "esg_pos",
                              "len_neg", "len_mix", "len_pos", "len_none",
                              "dpes_awe_total", "panas_n_total", "panas_p_total", "age", "sex_M")) +
  geom_errorbar(aes(xmin = coefficient - 1.96*se, xmax = coefficient + 1.96*se), width = .5) +
  geom_vline(xintercept = 0, linetype = "dotted") +
  theme_bw() + 
  theme(panel.border = element_blank(), panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(), axis.line = element_line(colour = "white"))


### ________________________________________________________ ####
### @@@ VI. MULTIVARIATE MACHINE LEARNING MODELING [FIG 3B-D] @@@ ####
## > VI-1. Setting for H2O Analysis ####
data.h2o <- data[, c("sex_M", "age", "panas_p_total", "panas_n_total", "dpes_awe_total",
                     "awes_avg", "esg_pos", "esg_neg", "esg_eci", "arousal", "motion_sickness",
                     "len_none", "len_pos", "len_mix", "len_neg")]

h2o.init()
data.h2o <- as.h2o(data.h2o)
h2o.describe(data.h2o)
colnames(data.h2o)

data.h2o.train <- h2o.splitFrame(data.h2o, ratios = 0.8, seed = 1)[[1]]
data.h2o.test <- h2o.splitFrame(data.h2o, ratios = 0.8, seed = 1)[[2]]
  
h2o.y <- "awes_avg" 
h2o.x <- c("sex_M", "age", "panas_p_total", "panas_n_total", "dpes_awe_total", 
           "esg_pos", "esg_neg", "esg_eci", "arousal", "motion_sickness", 
           "len_none", "len_pos", "len_mix", "len_neg")

## > VI-2. Model Fitting with 5-Fold Cross Validation ####
# ______________________________________________________________________________________________________________________
# [NOTE!] To reproduce the results, please load the pre-trained model before proceeding with the analysis. (line 448 ~ )
# ______________________________________________________________________________________________________________________
# automl.models.5cv <- h2o.automl(x = h2o.x, 
#                                 y = h2o.y, 
#                                 training_frame = data.h2o.train, 
#                                 max_models = 20, 
#                                 seed = 1,
#                                 nfolds = 5,
#                                 keep_cross_validation_fold_assignment = TRUE,
#                                 keep_cross_validation_predictions = TRUE,
#                                 keep_cross_validation_models = TRUE)

# automl.board <- h2o.get_leaderboard(automl.models.5cv, extra_columns = "ALL")
# automl.board.df <- as.data.frame(automl.board)


## > VI-3. Model Selection ####
# automl.best.model <- automl.board.df[3, "model_id"]
# automl.best.GBM <- h2o.getModel(automl.best.model)           # we chose 3rd model as the best model instead of the 1st and 2nd ensemble models
                                                               # because the Stacked Ensemble models don't provide feature importance info.
# h2o.saveModel(automl.best.GBM,                              
#               path = "../results/behavioral_results", force = TRUE)

# automl.base.model <- automl.board.df[14, "model_id"]         # GLM model
# automl.base.GLM <- h2o.getModel(automl.base.model)
# h2o.saveModel(automl.base.GLM,                               
#               path = "../results/behavioral_results", force = TRUE)
# ______________________________________________________________________________________________________________________

## > VI-4. Performance Evaluation (Fig 3B) #### 
automl.best.GBM <- h2o.loadModel("../results/behavioral_results/autoML_best-model_GBM")
automl.base.GLM <- h2o.loadModel("../results/behavioral_results/autoML_base-model_GLM")

automl.best.GBM.perf <- h2o.performance(automl.best.GBM, xval = T)
automl.base.GLM.perf <- h2o.performance(automl.base.GLM, xval = T)

res.h2o.perf.df <- data.frame(matrix(ncol = 3, nrow = 8))

colnames(res.h2o.perf.df) <- c("model", "metric", "value")
res.h2o.perf.df$model <- c(rep("GBM", 4), rep("GLM", 4))
res.h2o.perf.df$metric <- rep(c("MSE", "RMSE", "MAE", "R2"), 2)
res.h2o.perf.df$value <- c(automl.best.GBM.perf@metrics$MSE, automl.best.GBM.perf@metrics$RMSE, 
                           automl.best.GBM.perf@metrics$mae, automl.best.GBM.perf@metrics$r2,
                           automl.base.GLM.perf@metrics$MSE, automl.base.GLM.perf@metrics$RMSE, 
                           automl.base.GLM.perf@metrics$mae, automl.base.GLM.perf@metrics$r2)

res.h2o.perf.df$model <- as.factor(res.h2o.perf.df$model)
res.h2o.perf.df$metric <- as.factor(res.h2o.perf.df$metric)
res.h2o.perf.df$value <- as.numeric(res.h2o.perf.df$value)

write.csv(res.h2o.perf.df, "../results/behavioral_results/autoML_performance_table.csv")

# visualization
fig3b <- ggplot(res.h2o.perf.df, aes(x = metric, y = value, fill = model)) + 
  geom_col(position = "dodge") + 
  geom_text(aes(label = round(value, 3)), vjust = -0.50, color = "black", size = 3, 
            position = position_dodge(0.9)) +
  scale_fill_manual(values = c("GBM" = "#E1BEE7", "GLM" = "#BDBDBD")) +
  scale_x_discrete(limits = c('R2', 'MAE', 'MSE', 'RMSE')) +
  theme_bw() + 
  theme(panel.border = element_blank(), panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(), axis.line = element_line(colour = "black"))

## > VI-5. Feature Interpretation ####
# >> VI-5-a. variable importance (Fig 3C) ####
automl.best.GBM.varimp <- h2o.varimp(automl.best.GBM)
automl.best.GBM.varimp$variable <- reorder(automl.best.GBM.varimp$variable, 
                                          -automl.best.GBM.varimp$scaled_importance)

# visualization
fig3c <- ggplot(automl.best.GBM.varimp, aes(y = scaled_importance, x = variable)) +
  geom_bar(position = "dodge", stat = "identity", 
           aes(fill = ifelse(variable %in% variable[c(1, 4)], "#E1BEE7", "#BDBDBD")), 
           color = "white") +  
  geom_text(size = 2.50, aes(label = round(scaled_importance, 3)), 
            position = position_dodge(width = 0.90), vjust = 0.3, hjust = -0.2, angle = 90) +  
  theme_bw() + 
  theme(panel.border = element_blank(), panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(), axis.line = element_line(colour = "black"),
        axis.text.x = element_text(angle = 45, vjust = 1, hjust = 1)) +
  scale_fill_identity()

# >> VI-5-b. shapley value (Fig 3D) ####
fig3d <- h2o.shap_summary_plot(automl.best.GBM, data.h2o.test)

### ________________________________________________________ ####
### @@@ VII. INCLUDED vs. EXCLUDED IN EEG ANALYSIS (STable 1) @@@ ####
sub.included <- read.csv("../dataset/sub-list-latent_n27.csv", header = FALSE)
sub.included <- sub.included[[1]]
sub.excluded <- setdiff(sub.list.behav, sub.included)

df.included <- data[data$subjectkey %in% sub.included, ]
df.excluded <- data[data$subjectkey %in% sub.excluded, ]

## > VII-1. demographics - sex (male) ####
round((sum(df.included$sex_M)/length(df.included$sex_M)), 3)
round((sum(df.excluded$sex_M)/length(df.excluded$sex_M)), 3)

sex.bm.test <- brunnermunzel.test(df.included[df.included$clip == "SP", "sex_M"], 
                                  df.excluded[df.excluded$clip == "SP", "sex_M"])

## > VII-2. demographics - age ####
mean(df.included[df.included$clip == "SP", "age"])
sd(df.included[df.included$clip == "SP", "age"])
mean(df.excluded[df.excluded$clip == "SP", "age"])
sd(df.excluded[df.excluded$clip == "SP", "age"])

age.t.test <- t.test(df.included[df.included$clip == "SP", "age"], 
                     df.excluded[df.excluded$clip == "SP", "age"])

## > VII-3. awe-s ratings (SP) ####
mean(df.included[df.included$clip == "SP", "awes_avg"])
sd(df.included[df.included$clip == "SP", "awes_avg"])
mean(df.excluded[df.excluded$clip == "SP", "awes_avg"])
sd(df.excluded[df.excluded$clip == "SP", "awes_avg"])

awe.sp.t.test  <- t.test(df.included[df.included$clip == "SP", "awes_avg"], 
                         df.excluded[df.excluded$clip == "SP", "awes_avg"])

## > VII-4. awe-s ratings (CI) ####
mean(df.included[df.included$clip == "CI", "awes_avg"])
sd(df.included[df.included$clip == "CI", "awes_avg"])
mean(df.excluded[df.excluded$clip == "CI", "awes_avg"])
sd(df.excluded[df.excluded$clip == "CI", "awes_avg"])

awe.ci.t.test  <- t.test(df.included[df.included$clip == "CI", "awes_avg"], 
                         df.excluded[df.excluded$clip == "CI", "awes_avg"])

## > VII-5. awe-s ratings (MO) ####
mean(df.included[df.included$clip == "MO", "awes_avg"])
sd(df.included[df.included$clip == "MO", "awes_avg"])
mean(df.excluded[df.excluded$clip == "MO", "awes_avg"])
sd(df.excluded[df.excluded$clip == "MO", "awes_avg"])

awe.mo.t.test  <- t.test(df.included[df.included$clip == "MO", "awes_avg"], 
                         df.excluded[df.excluded$clip == "MO", "awes_avg"])

p.adjust(c(sex.bm.test$p.value, age.t.test$p.value, awe.sp.t.test$p.value, 
           awe.ci.t.test$p.value, awe.mo.t.test$p.value), method = "fdr")

### ________________________________________________________ ####
### RESULTS ####
fig2a.SP
fig2a.CI
fig2a.MO
fig2a.PA

fig2b

fig2c.dur
fig2c.int

fig3a
fig3b
fig3c
fig3d
